[CSL-1190] Get rid of DB/Iterator, cleanup db/
