+++ db/Pos/DB/GState/Balances.hs
@@ -19,11 +19,10 @@ import           Pos.Binary.Core      ()
 import           Pos.Core.Types       (Coin, StakeholderId)
 import           Pos.Util.Util        (maybeThrow)
 
-import           Pos.DB.Class         (MonadDBRead)
+import           Pos.DB.Class         (DBIteratorClass (..), MonadDBRead)
 import           Pos.DB.Error         (DBError (DBMalformed))
 import           Pos.DB.Functions     (encodeWithKeyPrefix)
 import           Pos.DB.GState.Common (gsGetBi)
-import           Pos.DB.Iterator      (DBIteratorClass (..))
 
 data BalanceIter
