--- db/Pos/DB/Iterator.hs
@@ -1,3 +1,0 @@
--- | Re-exports DB.Iterator
-
-{-# OPTIONS_GHC -F -pgmF autoexporter #-}
