+++ lrc/Pos/Lrc/Fts.hs
@@ -17,7 +17,6 @@ import           Pos.Core.Constants (epochSlots)
 import           Pos.Core.Types     (Coin, LocalSlotIndex (..), SharedSeed (..),
                                      SlotLeaders, StakeholderId, mkCoin)
 import           Pos.Crypto         (deterministic, randomNumber)
--- import           Pos.DB.Iterator    (MonadIterator (..))
 
 -- | Whereas 'Coin' stores an amount of coins, 'CoinIndex' is an identifier
 -- for a particular coin.
